requirements.txt
pettingzoo
gymnasium
numpy
pygame
pillow
torch
safetensors
